Changed GEMM test logic
It now works similarly to BLAS2 GEMV tests in that it uses a gtest
TEST_P test thing. This also covers different transposation values,
so they are no longer in seperate files.

In addition the batched/nonbatched versions are split up into
seperate files.

This also required changing assert_t_equal to a macro so it fails
on the first assert, rather than printing all of them.