Add temporary default constructor to BufferIterator
SYCL buffers do not need to be default constructible, but there is a
proposal to specify that they must be. It is useful to be able to
default construct `BufferIterator`s, as they can be used as a
placeholder and reassigned to a `BufferIterator` that contains a valid
SYCL buffer if needed.

As the buffer default constructor is not currently available in SYCL we
provide a temporary constructor that will construct a `BufferIterator`
but it would be a user error to pass this object into any SYCL-BLAS
function.

Once the default constructor is provided in the SYCL specification then
this temporary constructor can be removed.

See: https://github.com/codeplaysoftware/standards-proposals/blob/master/default-constructed-buffers/default-constructed-buffers.md