Only force inline functions on device
We want to always force inline all functions in the compute kernels, to
ensure that no intermediate structures are being forced to be
constructed that are not needed. However this tends to increase compile
times and there is no need to do this in host code as well. Instead we
can force inline only when compiling for the device and leave it up to
the host compiler whether it should inline host functions.