Fix reduction types in GEMM reduction (#203)
The input types for GEMM can be matrix views of const types (e.g.
`MatrixView<const float, ...>`, however the temporary `cube_buffer` used
in the GEMM reduction needs to have a non-const type (`MatrixView<float,
...>`), as it is the output buffer for the first step of the
computation. If the input type is used as part of the `Reduction` type
then this causes conflicting template types. Instead the actual type of
the temporary buffer should be used in the `Reduction` type.