Fix compile problems with GEMM auto tuners
Recent changes to the build system and GEMM kernels caused problems with
the auto tuners. Some small fixes are needed to make them compile.

For now we use a hardcoded vector size of 1, but we should look into
extending the tuners to try different vector sizes to get optimal
performance.