Refactor GEMM tests to use single function
Rather than having a duplicated verification function in each source
file, we can provide a single GEMM verification function in a header,
and then unify how the tests are registered in GTest with a single
macro. This allows different kernel parameters and test cases to be
easily added and amended.

Provide a cached SYCL queue to avoid recompiling kernels in each test
case. This can significantly improve test runtime.