Reorder cmake commands to prevent overwriting options (#221)
* Reorder cmake commands to prevent overwriting options

The SYCL-BLAS cmake scripts are very sensitive to the order in which
options are declared and used. Recently this causes link failures when
cmake is run with default options, as the `GEMM_TALL_SKINNY_SUPPORT`
option is used before it has been declared to have a default of `ON`.
This means that the first time the library object targets are generated
they do not include any tall skinny GEMM kernels, however when the host
launcher object is compiled the option has been defaulted on, so the
host expects these kernels to be available.