Add no local GEMM partial vectorization and selectable vectorization types (#230)
Adds the first vectorized version of the no_local GEMM kernel as a seperate file, also adds an enum to GEMM configurations to allow selecting the type of vectorization to use. This allows us more options for picking the best kernel configuration for different backends.

This partial vectorization of the no_local kernel is only vectorizable when A is not transposed and B is transposed.