Fix compile problem with GEMV
A user might want to provide a buffer of `const` values as an input of
GEMV, however the current kernel will infer that the value to use for all
inermediates should then be `const`. By removing the `const` from the type
we can use `const` inputs and have mutable intermediates.