Fix GEMV propagating NaNs when beta is zero
It is a common pattern to call routines like GEMM and GEMV with
uninitialized data in the output buffer when specifying beta to be zero.
This can cause problems if the uninitialized data happens to contain bit
patterns that correspond to NaN values, which can then get incorrectly
propagated to the output, as `0 * NaN == NaN` and not `0` as required.

To avoid this, provide an alternative final kernel that avoids loading
and multiplying the output by zero.