Add MiniBatchSize parameter
The parameter is only used in the ARM_GPU configuration. It allows each
work item to process a batch of size MiniBatchSize at a time. This
changes the access pattern of the matrices which may work well with the
caching system.

A MiniBatchSize of 1 should compile to the same kernel than before.