Refactor GEMM configurations
GEMM configurations are stored in a nested list.
However, CMake doesn't really support nested lists.
This patch removes the list of configurations
because it's not needed.
Instead, each configuration is added
using `add_gemm_configuration` function
which generates a file directly
and adds it the list of GEMM sources.

This doesn't require named variables anymore
and simplifies specifying configurations for different data types.
Each target has a list of supported data types
and configurations are generated for each type.