Initial quantization
In order to support different data types,
there has to be a way to convert data from `T*` to `float*`,
and vice-versa.
At minimum this is required to call reference BLAS,
which can only accept data as `float*` or `double*`.

This patch introduces that functionality
by providing a new public function `blas::_quantize`.
The function takes a policy handler,
a container of the input type,
and a container of the output type.
The containers are passed as SYCL buffers
to the `blas::internal::_quantize` function.

At the moment the only "quantization" provided
is a `static_cast` on each element,
which should work in most cases.
It is possible to specialize the quantization
by specializing `blas::internal::QuantizeKernel`
or even `blas::internal::Quantize`.

Important C++ files:
* `include/quantize.h` exposes the function publicly
* `src/quantize/quantize.hpp` provides internal functionality
* `src/quantize/quantize.cpp.in` provides a file template
  where templates for internal functionality are instantiated
  as needed by the data types.
  `float` and `double` don't need this file.
* `src/quantize/quantize_base.cpp.in` instantiates templates
  for `float` and `double`

Right now only the `asum` test was modified
to store the data as only `float*` and `double*`
and perform quantization for the tested type.
For this, special helper functions were provided.