Call quantized_copy_to_host in benchmarks
The patch that quantized all tests and benchmarks
ignored the fact that `make_quantized_buffer`
cannot copy back to the original data.
This would only be possible with non-quantized types,
whereas for others it creates another buffer of a different type.

This patch addresses this by adding the missing copy-back to benchmarks.
It also adds a Doxygen note about data not being copied back.