[Quantize] Use get_count to schedule quantization
When launching the quantization kernel,
the wrong number of work-items was specified to `parallel_for`,
leading to out of bounds memory access.
This didn't affect `float` and `double` types
because they don't need a quantization kernel.

The fix is to only schedule as many work-items
as there are data elements (using `get_count`).