Fix how libraries propagate in SYCL-BLAS lib
Instead of adding an absolute path to the interface include dirs,
we should use the PUBLIC link, since any SYCL-BLAS target will
transitively depend on ComputeCpp or some other SYCL implementation
as well (SYCL-BLAS uses SYCL constructs in its public headers).

Without this change, packages created by installing the project
will have absolute paths of the machine it was built on, making
the package unrelocatable. With it, a requisite of the SYCL
library is added to the targets. The SDK includes are still
propagated for the export file, but removed for installs, as the
headers are installed alongside the SYCL-BLAS headers so can be
found in the same location.