Add test parameters to GTest names
The tests are renamed with the parameters value.
This allows for a finer granularity than using STRESS_TESTING
when excluding tests. The parameters name are of the form
(__<paramName>_<paramValue>)*.

The tests are now run with an extra flag to produce an xml summary
when running ctest.

Also fix undefined variable that led to tests printing the help message.