Fix compilation error with hipSYCL (#447)
This PR resolves a compilation error that occurs when targeting NVIDIA GPUs and using hipSYCL as the SYCL compiler. Additionally, it re-enables benchmarking with hipSYCL.

It's important to note that the fix for this compiler issue with all targets will be complete once PR #434 is merged. This is because PR #434 includes essential safeguards to prevent compilation failures, which are necessary due to a missing feature in hipSYCL.