Update configuration for gemm on AMD GPUs
Following preliminary investigation and tuning with the auto tuner, these are the
new configurations for gemm that provide the best perfomance.
The selection of the configuration is now based on the arithmetic
intensity and not only on _N and _M dimension. A part from general gemm
no other implementation are affected.