[benchmark-rocblas] Add check for compiler-rt libs
Rocm library needs compiler-rt library in order to compile.
Not all distribution of clang provides it and it is not built
by default from LLVM. Since it is mandatory, users can always use a
version compatible with their version of clang or build it from source.
This solution supports both possibility.

Signed-off-by: nscipione <nicolo.scipione@codeplay.com>