Update rotmg implementation
Due to compatibility with oneMKL rotmg needs to handle scalar value not
passed as device pointer.
Moreover, a bug in the compiler make the event dependency not always
respected when using OpenCL backend on CPUs, so it is necessary a
work-around specifically for the default target which can be used for
any hardware.